feat: adopt tagpath consumer contract (p6tsi)
Bumps tsift to 0.1.47 and wires tagpath's `.naming/index.json` as a stable
symbol-graph adapter so `tsift search` can annotate hits with content-
addressable `mem:<sha256[0..16]>` handles surviving ordinary edits.

- New `src/tagpath_adapter.rs`: `try_load`, `TagpathAdapter`, `LoadResult`,
  `HandleResolution`. Auto-detects `.naming/index.json` at the project
  root, runs `tagpath::index::check` for freshness, and exposes
  `family_for_member` / `handle_for_member` / `lookup_handle` lookups.
- `index::SymbolHit` gains an optional `tagpath_handle` field
  (`#[serde(skip_serializing_if = "Option::is_none")]`).
- `cmd_search_with_budget` now calls `annotate_hits_with_tagpath` so the
  field is populated when the adapter loads. Stale indexes log a
  `tagpath_index_stale: true` stderr diagnostic and fall back to live
  extraction.
- New search flags: `--no-tagpath` (skip lookup) and `--tagpath-strict`
  (fail closed on a stale index).
- `Cargo.toml`: tagpath dep moved to local path with a slim feature set
  (`lang-rust,lang-python,lang-javascript,lang-typescript`, no `mcp`,
  no `watch`, no `dyn-grammar`); tree-sitter bumped to 0.26 to match.
- 7 new adapter tests cover missing / fresh / stale / family / handle /
  round-trip cases. Full suite 824 → 831 passing.
- SPEC `Tagpath integration` section + VERSIONS.md 0.1.47 entry document
  the surface; cross-links tagpath `SPEC.md` §15.

+++ src/lang/mod.rs
@@ -280,7 +280,7 @@ impl Lang {
                 let cmd = name_node.utf8_text(source).unwrap_or("");
                 if cmd == "alias" {
                     for i in 0..node.named_child_count() {
-                        if let Some(arg) = node.named_child(i)
+                        if let Some(arg) = node.named_child(i as u32)
                             && (arg.kind() == "concatenation" || arg.kind() == "word")
                         {
                             let text = arg.utf8_text(source).unwrap_or("");
